+++ cli/csstools-cli/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changes to CSSTools CLI
 
+### Unreleased (minor)
+
+- Added `@csstools/postcss-is-pseudo-class`
+
+```sh
+npx -y @csstools/cli postcss-is-pseudo-class
+```
+
 ### 1.0.0 (January 2, 2022)
 
 - Initial version